feat(icons): add new SVG icons
Add the following SVG icons to the project:
- arrow.svg: An arrow icon
- capa.svg: A cap icon
- chip.svg: A chip icon
- close.svg: A close icon
- facebook.svg: A Facebook icon
- insta.svg: An Instagram icon
- laptop.svg: A laptop icon
- leftarrow.svg: A left arrow icon
- menu.svg: A menu icon
- plane.svg: A plane icon
- rightarrow.svg: A right arrow icon

These icons were added to enhance the visual representation of the application and provide a better user experience.

feat(icons): add TikTok SVG icon

A new SVG icon for TikTok has been added to the project's assets. The icon is a black and white TikTok logo with a size of 24x24 pixels.